Dead-simple grammar that can read natural numbers [parsimonious]
It should be fairly clear what's going on in this grammar: numbers are
any string of text that contains 1 or more digits.

The Calculator class might not be obvious. Parsimonious defines a
NodeVisitor class that's used for walking the AST returned by
Grammar.parse. For each rule in your grammar, you define a method called
visit_<that_rule>. So for our simple grammar with just a rule called
`number`, we define a `visit_number` method on our AST walker. It gets a
"node", which is an object containing the text that the rule matched,
and a list of visited_children, which in this case is empty. We'll see
visited_children play a larger part later.