Parse decimal numbers [parsimonious]
Numbers are now a series of digits, possibly with a . in the middle of
them. This parser doesn't accept inputs like "15." or ".15", which are
arguably valid input. Personally I figure it won't kill the user to type
a 0 once in a while.

generic_visit might need some explanation. parsimonious's NodeVisitors
call generic_visit when the group they're matching doesn't have a
particular name. Rules like the `number` rule, which consists of several
parts, are implicitly transformed into a named rule and several
anonymous rules that're visited by generic_visit. We don't care about
the lower-level matches, just the outermost number, so generic_visit
doesn't need any functionality. However, the default implentation in
NodeVisitor raises a NotImplementedError, so we have to override it with
an empty method.