Now the parser can add two numbers [parsimonious]
The "plus" rule returns a function that takes two numbers and adds them
together. If you're not familiar with higher-order functions you might
be surprised to see a `return lambda...`, but in fact it fits quite
nicely with what we're doing. In the language we're defining, `+` is a
function of two numbers, so our interpreter transforms the literal
string '+' into the very function it's meant to represent.

The "addition" rule finally makes use of that visited_children argument
we've been ignoring for so long. Looking up at the grammar, we see that
an addition is a number, a plus, and another number. Since the
addition rule refers to other rules, its visited_children is a list of
the return values of the visit methods for the referred-to rules. In
this case, it's a float returned by visit_number, a function returned by
visit_plus, and another float from visit_number. We grab the function
from the middle of the list and call it on the left and right numbers.

visit_expression also uses its visited_children argument, but in this
case it doesn't do much with it. The `expression` rule simply expresses
that you can have either a number or an addition at the topmost level,
so by the time we visit its node, all the work's already been done. We
just keep returning the value calculated in the lower-level rules.

We could simplify our grammar somewhat:
    expression = (number plus number) / number
    plus       = "+"
    number     = "-"? ~"[0-9]+" ("." ~"[0-9]+")?
But this would result in a significantly-more-complicated implementation
in visit_expression.