Allow chaining additions [parsimonious]
Now we have some recursion in our grammar! The `expression` refers to
the `addition` rule, which in turn refers back to the `expression` rule.
Anchoring the `addition` rule with the non-recursing `number` rule on
the left ensures our recursion has a terminus. If we instead define the
`addition` rule as `expression plus expression`, parsimonious runs into
Python's recursion depth limit when trying to parse the input text.