Support various operators [parsimonious]
Note how visit_operation, other than being renamed, is no different from
visit_addition. This is a consequence of making visit_operator (nee
visit_plus) a higher-order function. See, functional programming is
great!

Also take note of the "operators are right-associative" test. This is a
consequence of the decision to define the `operation` rule as `number
operator expression`: The right-hand side of the operator is an
expression, and its evaluated result is what's available to the
operator. It is somewhat harder in Parsimonious to write a
left-associative language.

In the majority of cases, the particular flavor of parser generator you
use isn't too important. This is a case where it matters: parsimonious
generates PEG parsers, for which left-associativity is possible, but
harder, to define. Some parser generators generate LL or LR parsers; the
second letter indicates which "derivation" of the language the parser
uses. An LL parser will tend to generate left-associative languages. An
LR parser will tend to generate right-associative languages. For various
reasons, LR parsers are much more common than LL; thus most programming
languages (including Python) are right-associative.