Add parentheticals to our language [parsimonious]
Note that the left side of the `operation` rule can now self-recurse:
operation->element->parenthetical->expression->operation. However, a
parenthetical must have...a parenthesis, so we're still anchored and
don't endlessly recurse.

Look at how everything that referred directly to the `number` rule now
refers to `element`. This sort of composition is important in keeping a
grammar clean. We could do away with the `element` rule:
    expression    = operation / parenthetical / number
    operation     = (parenthetical / number) operator expression
But now we have repetition. Just like in any other programming task,
when you see repetition in your grammar you should extract it into a
single place to which you can make reference.

(also, we'd have to change generic_visit if we didn't have a named rule
for elements.)