Now the parser can add numbers together [ply]
This is not quite the design recommended by the ply documentation, but I
think it's a nice clean design. The tokenizer returns the values
associated with each token, and the parser consumes those values. In
this case, the value associated with '+' happens to be a function of two
numbers, so that's what we'll return.

If you've already read the parsimonious version of this parser, you'll
notice that PLY allows `expression: expression plus expression` whereas
parsimonious would've gone into an infinite loop on that grammar. This
is a distinction between the LALR grammars used by PLY and the PEG
grammars used by parsimonious.

Note that this uncovered a bug in my decimal-numbers regex. Oops! I
blame PLY for being so regex-oriented. It's definitely not my fault for
making a classic regex mistake. Nope.