Add various operators [ply]
Now you can see what I meant about a nice clean design from returning
functions in the tokenizer. Each operator token returns the associated
function, and the parser just calls it. In a more classic yacc
implementation, the operator tokens would just return their string
representations, and you'd have a big if-elsif-elsif-elsif block where
you look up the behavior associated with the strings.

You may notice that the number rule changed here. Previously a unary '-'
was considered part of the numeric token, whereas now it's a separate
operator. It turns out that with two different ways to interpret the '-'
token, you end up with an ambiguous grammar and the behavior is
unpredictable. Having the unary '-' pass an explicit 0 to the binary
subtraction function is either an elegant workaround or a kludgy hack,
depending on your viewpoint.