Add parentheticals to the language [ply]
The grammar here is pretty straightforward; an expression can now be an
otherwise-valid expression surrounded by parentheses.

You'll notice that we had to hack in parse-error expressions. By
default, PLY prints parse errors to STDOUT and returns None. That's not
very useful in any context I can think of. Instead, we inject a logger
that blackholes all input except calls to error(); calls to error()
raise a ParseError. It's pretty kludgey, but as the tests show, it at
least works.