Dead-simple grammar that can read natural numbers [parsley]
Our grammar has a single rule, "number." The syntax here is a little
more than just a grammar: Parsley supplies a few operators that make it
easier to build a parser and grammar in one go. the angle brackets in
`<digit+>` mean that the `digit+` rule returns a single string
containing all the digits it matched, rather than a list. The `:ds`
assigns that matched group to a locally-available variable called `ds`.
Finally, everything after the `->` is python code, the return value of
which will be the ultimate return value of the rule.

Tying it all together, `<digit+>:ds -> int(ds)` matches one or more
numerals and casts them to a Python int.

(note that the `digit` rule is supplied out-of-the-box by Parsley)