Now our parser can add two numbers [parsley]
Notice that the `+` rule returns a function. This seems a little
overcomplicated now, but when we start adding operators you'll see that
it provides a certain elegance.

Notice also that we had to change the function called by calculate().
The Parsley function makeGrammar returns a function that takes a piece
of text and returns a parser; that parser has functions with names
corresponding to the rules defined in the grammar. Since our grammar's
outermost rule is now expression, that's the function we should call.