Allow chaining additions [parsley]
Notice that our addition rule requires a number on the left side. This
is one of the cases where we do care about the class of grammar
generated by our parser-generator. Specifically, Parsley generates a PEG
grammar, which means that it breaks if you define a self-recursive rule.
If our `addition` rule were `expression plus expression`, the left-side
`expression` term would refer back to `addition`, creating a
non-terminating situation.