Add various operators [parsley]
Now you can see the elegance of having our operator rules return
functions: the `operation` rule simply needs to apply the returned
function to its left- and right-hand arguments, leaving the
symbol-to-operation logic to the operator rules.

The `roll` function is fairly compactly-written, but if you read it
carefully you should be able to see what it does: generate `num` random
integers between 1 and `size`, then add them all up.

This commit finally reveals the import of that dict we've been passing
to makeGrammar: it is a mapping of names to values that will be
available inside our grammar.