Support for adding two numbers [funcparserlib]
The `skip` combinator forces its supplied parser to match, but doesn't
append the matched string to the list of results. This is perfect for
our needs: we just pass the numeric parts of the expression to the
built-in `sum` function, which takes a list of numbers and adds them
all together.